requirements.txt
alembic==1.13.3
astroid==3.3.5
atomicwrites==1.4.1
attrs==24.2.0
bidict==0.23.1
black==24.10.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
dnspython==2.7.0
dotmap==1.3.30
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Migrate==4.0.7
Flask-SocketIO==5.4.1
Flask-SQLAlchemy==3.1.1
future==1.0.0
greenlet==3.1.1
importlib-metadata==8.5.0
iniconfig==2.0.0
isort==5.13.2
itsdangerous==2.2.0
Jinja2==3.1.4
lazy-object-proxy==1.10.0
Mako==1.3.6
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
netifaces==0.11.0
packaging==24.1
pathspec==0.12.1
pillow<11,>=7.1.0
pip==24.2
platformdirs==4.3.6
pluggy==1.5.0
py==1.11.0
py-expression-eval==0.3.14
pycparser==2.22
pylint==3.3.1
pylint-flask==0.6
pylint-flask-sqlalchemy==0.2.0
pylint-plugin-utils==0.8.2
pyparsing==3.2.0
pyserial==3.5
pytest==8.3.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-editor==1.0.4
python-engineio==4.10.1
python-socketio==5.11.4
PyYAML==6.0.2
setuptools==75.2.0
six==1.16.0
SQLAlchemy==2.0.36
toml==0.10.2
tomli==2.0.2
typed-ast==1.5.5
typing_extensions==4.12.2
watchdog==5.0.3
Werkzeug==3.0.4
wheel==0.44.0
wrapt==1.16.0
zipp==3.20.2
zope.event==5.0
zope.interface==7.1.0